+++ src/__tests__/config.test.ts
@@ -34,6 +34,7 @@ describe('config.ts', () => {
         "cloudUrl": "file://~",
         "command": "up",
         "commentOnPr": false,
+        "configMap": undefined,
         "githubToken": "n/a",
         "isPullRequest": false,
         "options": Object {
@@ -95,6 +96,7 @@ describe('config.ts', () => {
         "cloudUrl": "file://~",
         "command": "up",
         "commentOnPr": true,
+        "configMap": undefined,
         "githubToken": "n/a",
         "isPullRequest": false,
         "options": Object {
@@ -148,6 +150,7 @@ describe('config.ts', () => {
         "cloudUrl": "file://~",
         "command": "up",
         "commentOnPr": false,
+        "configMap": undefined,
         "githubToken": "n/a",
         "isPullRequest": true,
         "options": Object {