+++ src/config.ts
@@ -42,6 +42,7 @@ export const config = rt
     rt.Partial({
       // Optional options
       cloudUrl: rt.String,
+      configMap: rt.String,
       githubToken: rt.String,
       upsert: rt.Boolean,
       refresh: rt.Boolean,
@@ -62,6 +63,7 @@ export async function makeConfig(): Promise<Config> {
     commentOnPr: parseBoolean(getInput('comment-on-pr')),
     upsert: parseBoolean(getInput('upsert')),
     refresh: parseBoolean(getInput('refresh')),
+    configMap: getInput('configMap'),
     isPullRequest: context?.payload?.pull_request !== undefined,
     options: {
       parallel: parseNumber(getInput('parallel')),